feat(fcm): Enable disabling of XMPP entirely using properties
- If upstream is disabled, then the XMPP downstream sender can also be not used.
