+++ src/main/java/org/radarbase/fcm/upstream/error/LoggingErrorHandlingStrategy.java
@@ -25,10 +25,12 @@ import com.fasterxml.jackson.databind.JsonNode;
 import lombok.extern.slf4j.Slf4j;
 import org.radarbase.fcm.config.ReconnectionEnabledXmppConnectionFactoryBean;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
+@ConditionalOnProperty(name = "fcmserver.xmpp.upstream.enable", havingValue = "true")
 public class LoggingErrorHandlingStrategy implements ErrorHandlingStrategy {
 
   @Autowired private transient ReconnectionEnabledXmppConnectionFactoryBean connectionFactoryBean;
