+++ src/main/resources/application-dev.properties
@@ -71,6 +71,7 @@ fcmserver.senderid=1043784930865
 fcmserver.serverkey=AAAA8wZuFjE:APA91bHmsDE1mhi1aQrmcTNT--RAofoVdcgc6wkCS1BTeHui4mdwOhaeum_D4dMcaQ4pfe9vuwJWZFWOF9EJ-JN8mkzMZ2Crm7LGK1WD2NMA-xd3xIkoNQx8jU0dySAzhbzWyI5LB9ljvliUgXWVFaj3-uQSZfzfdA
 fcmserver.fcmsender=org.radarbase.fcm.downstream.AdminSdkFcmSender
 fcmserver.host=fcm-xmpp.googleapis.com
+fcmserver.xmpp.upstream.enable=false
 # 5236 for test
 fcmserver.port=5236
 # radar-armt-appserver-test firebase
