+++ src/main/resources/application-prod.properties
@@ -28,6 +28,7 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/radar
 # jdbc:hsqldb:mem:/appserver for dev testing
 #jdbc:hsqldb:hsql://localhost:9001/appserver for running hsql separately in dev or testing
 spring.jpa.properties.hibernate.jdbc.time_zone=UTC
+spring.jpa.properties.hibernate.order_inserts=true
 ##spring.jpa.properties.hibernate.dialect=liquibase.ext.hibernate.database.HibernateGenericDialect
 
 # H2 BROWSER CONSOLE
@@ -64,7 +65,9 @@ radar.questionnaire.protocol.github.branch=master
 # Configure the following for production
 #fcmserver.senderid=
 #fcmserver.serverkey=
-fcmserver.fcmsender=org.radarbase.fcm.downstream.XmppFcmSender
+fcmserver.fcmsender=org.radarbase.fcm.downstream.AdminSdkFcmSender
+  #org.radarbase.fcm.downstream.XmppFcmSender
+fcmserver.xmpp.upstream.enable=false
 fcmserver.host=fcm-xmpp.googleapis.com
 # 5236 for test
 fcmserver.port=5235
